refactor: consolidate types, externalize prompts, and improve video service modularity
- Extract FFmpeg utilities into a dedicated helper module
- Move shared types into a single source to remove circular dependencies
- Externalize DeepSeek prompt templates for easier maintenance
- Clean up imports and standardize module structure across the video, workflow, and DeepSeek services ( well this last part msg is just english )
