+++ src/services/video.ts
@@ -10,84 +10,7 @@ import { writeFile, unlink } from "node:fs/promises";
 import { existsSync } from "node:fs";
 import * as logger from "../logger.ts";
 import { generateCaptions } from "./captions.ts";
-
-/**
- * Pan direction for vertical pan effect
- */
-type PanDirection = "up" | "down";
-
-/**
- * Pan parameters for zoompan filter
- */
-interface PanParams {
-  enabled: boolean;
-  direction: PanDirection;
-  yStart: number; // Starting Y position (pixels)
-  yEnd: number; // Ending Y position (pixels)
-}
-
-/**
- * Calculate pan parameters based on image aspect ratio and duration
- * @param duration - Scene duration in seconds
- * @returns Pan parameters for zoompan filter
- */
-function calculatePanParams(duration: number): PanParams {
-  // If pan effect is disabled, return disabled params
-  if (!PAN_EFFECT) {
-    return {
-      enabled: false,
-      direction: "down",
-      yStart: 0,
-      yEnd: 0,
-    };
-  }
-
-  // Target video dimensions
-  const VIDEO_WIDTH = 1920;
-  const VIDEO_HEIGHT = 1080;
-
-  // AI-generated image dimensions (4:3 aspect ratio)
-  const IMAGE_WIDTH = 1472;
-  const IMAGE_HEIGHT = 1104;
-
-  // Calculate scaled dimensions when fitting image to video width
-  // The image will be scaled to fit 1920px width while maintaining aspect ratio
-  const scaledHeight = (IMAGE_HEIGHT * VIDEO_WIDTH) / IMAGE_WIDTH; // = 1440px
-
-  // Calculate total vertical headroom (extra space above and below)
-  const totalHeadroom = scaledHeight - VIDEO_HEIGHT; // = 1440 - 1080 = 360px
-
-  // Use 30% of available headroom for visible pan effect
-  // NOTE: Increased from 15% to 30% to make pan more visible
-  const usableHeadroom = totalHeadroom * 0.30; // 30% of 360px = 108px
-
-  // Leave buffer zones at top and bottom (remaining 70% of headroom)
-  const bufferZone = (totalHeadroom - usableHeadroom) / 2; // = (360 - 108) / 2 = 126px
-
-  // Randomly choose pan direction (up or down)
-  const direction: PanDirection = Math.random() > 0.5 ? "down" : "up";
-
-  // Calculate start and end Y positions in pixels
-  let yStart: number;
-  let yEnd: number;
-
-  if (direction === "down") {
-    // Pan down: start at top buffer, end at top buffer + usable headroom
-    yStart = bufferZone;
-    yEnd = bufferZone + usableHeadroom;
-  } else {
-    // Pan up: start at top buffer + usable headroom, end at top buffer
-    yStart = bufferZone + usableHeadroom;
-    yEnd = bufferZone;
-  }
-
-  return {
-    enabled: true,
-    direction,
-    yStart: Math.round(yStart),
-    yEnd: Math.round(yEnd),
-  };
-}
+import { createFilterComplex } from "../utils/ffmpeg.ts";
 
 /**
  * Generate video from images and audio using FFmpeg with chunked rendering
@@ -154,87 +77,6 @@ export async function generateVideo(
     videoPath: outputPath,
     duration: totalDuration,
   };
-}
-
-/**
- * Create FFmpeg filter complex for image transitions
- * @param images - Sorted array of images with timing
- * @returns Filter complex string and total duration
- */
-function createFilterComplex(
-  images: DownloadedImage[]
-): { filterComplex: string; totalDuration: number } {
-  const filters: string[] = [];
-  let totalDuration = 0;
-
-  const imagesLength = images.length;
-
-  // Process each image with optional pan effect
-  for (let i = 0; i < imagesLength; i++) {
-    const image = images[i];
-    if (!image) continue;
-
-    const duration = (image.end - image.start) / 1000; // Convert ms to seconds
-    totalDuration += duration;
-
-    // Calculate pan parameters for this image
-    const panParams = calculatePanParams(duration);
-
-    if (panParams.enabled) {
-      // Apply pan effect using scale + crop (NO ZOOMPAN!)
-      //
-      // Why not zoompan?
-      // - zoompan is designed for zoom effects, not simple panning
-      // - It has complex frame timing issues with -loop 1
-      // - For vertical pan only, we just need: scale → crop with animated Y position
-      //
-      // Filter chain:
-      // 1. scale=1920:-1 → Scale to 1920px width, maintain aspect ratio (creates 1920×1440 for 4:3 images)
-      // 2. fps=30 → Set frame rate to 30fps BEFORE crop (ensures proper frame generation)
-      // 3. crop → Crop to 1920×1080 with animated Y position (this creates the pan effect)
-      // 4. setsar=1 → Set sample aspect ratio to 1:1
-      // 5. format=yuv420p → Convert to YUV420P color format
-
-      const fps = 30;
-      const totalFrames = Math.round(duration * fps);
-
-      // Animated Y position for crop filter
-      //
-      // The crop filter's y parameter can use expressions with 'n' (frame number)
-      // Formula: yStart + (yEnd - yStart) * (n / totalFrames)
-      //
-      // 'n' starts at 0 and increments by 1 for each frame
-      // We clamp it to totalFrames to prevent overshooting
-      const yExpression = `if(lte(n,${totalFrames}),${panParams.yStart}+(${panParams.yEnd}-${panParams.yStart})*n/${totalFrames},${panParams.yEnd})`;
-
-      // Crop filter parameters:
-      // - w=1920: Output width (crop to 1920px)
-      // - h=1080: Output height (crop to 1080px)
-      // - x=0: Horizontal position (no horizontal pan, start at left edge)
-      // - y=...: Vertical position (animated from yStart to yEnd)
-      //
-      // This crops a 1920×1080 window from the 1920×1440 scaled image,
-      // with the Y position animating from yStart to yEnd over totalFrames frames
-      filters.push(
-        `[${i}:v]scale=1920:-1,fps=${fps},crop=w=1920:h=1080:x=0:y='${yExpression}',setsar=1,format=yuv420p[v${i}]`
-      );
-
-      logger.debug("Video", `Image ${i + 1}: Pan ${panParams.direction} (${panParams.yStart}px → ${panParams.yEnd}px) over ${duration.toFixed(2)}s (${totalFrames} frames)`);
-    } else {
-      // No pan effect - use static image with scale and pad
-      filters.push(
-        `[${i}:v]scale=1920:1080:force_original_aspect_ratio=decrease,pad=1920:1080:(ow-iw)/2:(oh-ih)/2,setsar=1,fps=30,format=yuv420p[v${i}]`
-      );
-    }
-  }
-
-  // Concatenate all video segments
-  const concatInputs = Array.from({ length: imagesLength }, (_, i) => `[v${i}]`).join("");
-  filters.push(`${concatInputs}concat=n=${imagesLength}:v=1:a=0[outv]`);
-
-  const filterComplex = filters.join(";");
-
-  return { filterComplex, totalDuration };
 }
 
 /**
@@ -450,7 +292,7 @@ async function renderVideoInChunks(
     try {
       await unlink(chunkPath);
       logger.debug("Video", `Deleted ${chunkPath}`);
-      
+
       const assPath = chunkPath.replace('.mp4', '.ass').replace('chunk_', 'captions_');
       if (existsSync(assPath)) {
         await unlink(assPath);
@@ -673,69 +515,6 @@ async function concatenateChunks(
     ffmpeg.on("error", (error) => {
       logger.error("Video", `Failed to start FFmpeg concatenation: ${error.message}`);
       reject(new Error(`Failed to start FFmpeg concatenation: ${error.message}`));
-    });
-  });
-}
-
-/**
- * Add subtitles to an existing video file
- * @param inputVideoPath - Path to input video file
- * @param assFilePath - Path to ASS subtitle file
- * @param outputVideoPath - Path to output video file with subtitles
- */
-async function addSubtitlesToVideo(
-  inputVideoPath: string,
-  assFilePath: string,
-  outputVideoPath: string
-): Promise<void> {
-  return new Promise((resolve, reject) => {
-    // Escape the ASS file path for Windows (replace backslashes with forward slashes and escape colons)
-    const escapedAssPath = assFilePath.replace(/\\/g, "/").replace(/:/g, "\\:");
-
-    const ffmpegArgs = [
-      "-i",
-      inputVideoPath,
-      "-vf",
-      // NOTE: Do NOT use force_style parameter - it overrides inline styling tags in the ASS file
-      // The ASS file already contains all the styling information for word-by-word highlighting
-      `subtitles='${escapedAssPath}'`,
-      "-c:a",
-      "copy", // Copy audio without re-encoding
-      "-c:v",
-      "libx264",
-      "-preset",
-      "veryfast",
-      "-crf",
-      "23",
-      "-y",
-      outputVideoPath,
-    ];
-
-    logger.debug("Video", `Adding subtitles: ffmpeg ${ffmpegArgs.join(" ")}`);
-
-    const ffmpeg = spawn("ffmpeg", ffmpegArgs);
-
-    let stderrOutput = "";
-
-    ffmpeg.stderr.on("data", (data) => {
-      const output = data.toString();
-      stderrOutput += output;
-    });
-
-    ffmpeg.on("close", (code) => {
-      if (code === 0) {
-        logger.success("Video", "Subtitles added successfully");
-        resolve();
-      } else {
-        logger.error("Video", `Failed to add subtitles (exit code ${code})`);
-        logger.debug("Video", `FFmpeg stderr:\n${stderrOutput}`);
-        reject(new Error(`FFmpeg subtitle overlay exited with code ${code}`));
-      }
-    });
-
-    ffmpeg.on("error", (error) => {
-      logger.error("Video", `Failed to start FFmpeg for subtitles: ${error.message}`);
-      reject(new Error(`Failed to start FFmpeg for subtitles: ${error.message}`));
     });
   });
 }
