+++ src/types.ts
@@ -118,6 +118,23 @@ export interface CleanupResult {
   totalSize: number; // in bytes
 }
 
+// FFmpeg Types
+export type PanDirection = "up" | "down";
+
+export interface PanParams {
+  enabled: boolean;
+  direction: PanDirection;
+  yStart: number; // Starting Y position (pixels)
+  yEnd: number; // Ending Y position (pixels)
+}
+
+// Workflow Types
+export interface WorkflowResult {
+  videoPath: string;
+  duration: number;
+  minioUpload?: MinIOUploadResult;
+}
+
 // Progress Tracking Types
 export interface ProgressUpdate {
   step: string;
